1 The Physical/Guess Point Level 1 (Stateless)/2 - Arrange-Act-Assert Backwards/2.2 - Stats Calculator/src/index.ts
export function statsCalculator(numbers: number[]) {
  const result = {
    averageValue: 0,
    elementsCount: numbers.length,
    minValue: +Infinity,
    maxValue: -Infinity,
  };

  let numbersSum = 0;

  for (const num of numbers) {
    if (num < result.minValue) {
      result.minValue = num;
    }

    if (num > result.maxValue) {
      result.maxValue = num;
    }

    numbersSum += num;
  }

  result.averageValue = Number((numbersSum / result.elementsCount).toFixed(12));

  return result;
}
